Put RecentChange to own file and removed as inner class from RecentChangesFetcher
